requirements.txt
absl-py==2.1.0
accelerate==0.33.0
ai_tools==0.3.9
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
axial_positional_embedding==0.2.1
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip @ file:///nvme/scratch/aaradhye/CLIP
cognitive-face==1.5.0
contourpy==1.1.1
cycler==0.12.1
dalle-pytorch==1.6.6
datasets==2.21.0
decorator==5.1.1
diffusers @ git+https://github.com/huggingface/diffusers.git@4cfb2164fb05d54dd594373b4bd1fbb101fef70c
dill==0.3.8
dlib==19.24.6
docker-pycreds==0.4.0
einops==0.8.0
executing==2.0.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gast==0.4.0
gdown==5.2.0
git-lfs==1.6
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
huggingface-hub==0.24.7
idna==3.7
imageio==2.35.0
importlib_metadata==8.2.0
importlib_resources==6.4.2
inquirerpy==0.3.4
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
-e git+https://github.com/CompVis/latent-diffusion.git@a506df5756472e2ebaf9078affdde2c4f1502cd4#egg=latent_diffusion
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.6
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
numpy==1.24.3
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
peft==0.7.1
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rotary-embedding-torch==0.6.4
rsa==4.9
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
selective-optimizers==0.2
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
taming-transformers-rom1504==0.0.6
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics==1.4.1
torchvision==0.13.1+cu113
tqdm==4.66.5
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@8bd2b1e8c23234cd607ca8d63f53c1edfea27462
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.7
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
zprint==0.0.11
